demo/facechain_agent/requirements.txt
accelerate
transformers
diffusers
onnxruntime
modelscope
Pillow
opencv-python
torchvision
mmdet==2.26.0
mmengine
numpy==1.22.0
protobuf==3.20.1
timm
scikit-image
gradio
controlnet_aux==0.0.6
mediapipe
python-slugify
dashscope
datasets>=2.8.0
ipython
langchain
modelscope>=1.7.0
moviepy
ms-swift
openai
opencv-python
Pillow
pydantic==1.10.8
pytest
python-dotenv
soundfile
transformers>=4.29.0
transformers_stream_generator
